fix missing packages update
