+++ http/handler.go
@@ -14,7 +14,7 @@ const (
 )
 
 // HandleHealthJSON returns an HandlerFunc that can be used as an endpoints that exposes the service health
-func HandleHealthJSON(h health.Health) http.HandlerFunc {
+func HandleHealthJSON(h gosundheit.Health) http.HandlerFunc {
 	return func(w http.ResponseWriter, request *http.Request) {
 		results, healthy := h.Results()
 		w.Header().Set("Content-Type", "application/json")
